pyproject.toml
[tool.poetry]
name = "vgon"
version = "0.1.58"
description = ""
authors = ["tt <tt@nomail.mail>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tensorflow-cpu = "^2.16.2"
tqdm = "^4.66.4"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
